dkg2 package renamed to dkg
dkg2 was a temporary name when we were developing a new version of DKG
protocol while having the old dkg version still in place. Since we have
fully switched to the new version, we can do the rename.
