+++ Dockerfile
@@ -76,7 +76,7 @@ COPY ./pkg/net/gen $APP_DIR/pkg/net/gen
 COPY ./pkg/chain/gen $APP_DIR/pkg/chain/gen
 COPY ./pkg/beacon/relay/thresholdsignature/gen $APP_DIR/pkg/beacon/relay/thresholdsignature/gen
 COPY ./pkg/beacon/relay/gjkr/gen $APP_DIR/pkg/beacon/relay/gjkr/gen
-COPY ./pkg/beacon/relay/dkg2/result/gen $APP_DIR/pkg/beacon/relay/dkg2/result/gen
+COPY ./pkg/beacon/relay/dkg/result/gen $APP_DIR/pkg/beacon/relay/dkg/result/gen
 RUN go generate ./.../gen 
 
 COPY ./ $APP_DIR/
