+++ pkg/beacon/relay/dkg/dkg.go
@@ -1,11 +1,11 @@
-package dkg2
+package dkg
 
 import (
 	"fmt"
 	"math/big"
 
 	relayChain "github.com/keep-network/keep-core/pkg/beacon/relay/chain"
-	dkgResult "github.com/keep-network/keep-core/pkg/beacon/relay/dkg2/result"
+	dkgResult "github.com/keep-network/keep-core/pkg/beacon/relay/dkg/result"
 	"github.com/keep-network/keep-core/pkg/beacon/relay/gjkr"
 	"github.com/keep-network/keep-core/pkg/beacon/relay/group"
 	"github.com/keep-network/keep-core/pkg/chain"
